refactor(nx-governance): Separate extension runtime contracts from host-owned discovery and module loading
