+++ packages/governance/src/extensions/host.ts
@@ -1,643 +1,1 @@
-import { GovernanceWorkspace, Measurement, Violation } from '../core/index.js';
-import { GovernanceSignal } from '../signal-engine/index.js';
-import {
-  GovernanceExtensionDefinition,
-  GovernanceExtensionHost as GovernanceExtensionHostContract,
-  GovernanceExtensionHostContext,
-  GovernanceMetricProvider,
-  GovernanceMetricProviderInput,
-  GovernanceExtensionRulePack,
-  GovernanceRulePackInput,
-  GovernanceSignalProvider,
-  GovernanceSignalProviderInput,
-  GovernanceWorkspaceEnricher,
-  GovernanceWorkspaceEnricherInput,
-} from './contracts.js';
-import type { GovernanceExtensionDiagnostic } from './diagnostics.js';
-
-interface RegisteredGovernanceContribution<T> {
-  pluginId: string;
-  contribution: T;
-}
-
-export interface GovernanceExtensionRegistry {
-  metricProviders: RegisteredGovernanceContribution<GovernanceMetricProvider>[];
-  signalProviders: RegisteredGovernanceContribution<GovernanceSignalProvider>[];
-  rulePacks: RegisteredGovernanceContribution<GovernanceExtensionRulePack>[];
-  enrichers: RegisteredGovernanceContribution<GovernanceWorkspaceEnricher>[];
-}
-
-interface DiscoveredGovernanceExtension {
-  sourceSpecifier: string;
-  moduleSpecifier: string;
-  source: 'explicit' | 'legacy';
-  optional?: boolean;
-  definition: GovernanceExtensionDefinition;
-}
-
-export interface GovernanceExtensionLoadRequest {
-  sourceSpecifier: string;
-  moduleSpecifier: string;
-  source: 'explicit' | 'legacy';
-  optional?: boolean;
-}
-
-export interface DiscoverGovernanceExtensionsOptions {
-  loadRequests?: readonly GovernanceExtensionLoadRequest[];
-  moduleLoader?: GovernanceExtensionModuleLoader;
-}
-
-export type RegisterGovernanceExtensionsOptions =
-  DiscoverGovernanceExtensionsOptions;
-
-export type GovernanceExtensionModuleLoader = (
-  specifier: string
-) => Promise<unknown>;
-
-export interface GovernanceExtensionRegistrationResult {
-  registry: GovernanceExtensionRegistry;
-  diagnostics: GovernanceExtensionDiagnostic[];
-}
-
-export class GovernanceExtensionRegistrationError extends Error {
-  readonly diagnostics: GovernanceExtensionDiagnostic[];
-
-  constructor(message: string, diagnostics: GovernanceExtensionDiagnostic[]) {
-    super(message);
-    this.name = 'GovernanceExtensionRegistrationError';
-    this.diagnostics = diagnostics;
-  }
-}
-
-export class GovernanceExtensionHost
-  implements GovernanceExtensionHostContract
-{
-  readonly context: GovernanceExtensionHostContext;
-
-  private readonly registry: GovernanceExtensionRegistry = {
-    metricProviders: [],
-    signalProviders: [],
-    rulePacks: [],
-    enrichers: [],
-  };
-
-  private readonly pluginId: string;
-
-  constructor(context: GovernanceExtensionHostContext, pluginId: string) {
-    this.context = Object.freeze({
-      ...context,
-      options: Object.freeze({ ...context.options }),
-    });
-    this.pluginId = pluginId;
-  }
-
-  registerMetricProvider(metricProvider: GovernanceMetricProvider): void {
-    this.registry.metricProviders.push({
-      pluginId: this.pluginId,
-      contribution: metricProvider,
-    });
-  }
-
-  registerSignalProvider(signalProvider: GovernanceSignalProvider): void {
-    this.registry.signalProviders.push({
-      pluginId: this.pluginId,
-      contribution: signalProvider,
-    });
-  }
-
-  registerRulePack(rulePack: GovernanceExtensionRulePack): void {
-    this.registry.rulePacks.push({
-      pluginId: this.pluginId,
-      contribution: rulePack,
-    });
-  }
-
-  registerEnricher(enricher: GovernanceWorkspaceEnricher): void {
-    this.registry.enrichers.push({
-      pluginId: this.pluginId,
-      contribution: enricher,
-    });
-  }
-
-  toRegistry(): GovernanceExtensionRegistry {
-    return {
-      metricProviders: [...this.registry.metricProviders],
-      signalProviders: [...this.registry.signalProviders],
-      rulePacks: [...this.registry.rulePacks],
-      enrichers: [...this.registry.enrichers],
-    };
-  }
-}
-
-export async function discoverGovernanceExtensions(
-  options: DiscoverGovernanceExtensionsOptions = {}
-): Promise<DiscoveredGovernanceExtension[]> {
-  const result = await discoverGovernanceExtensionsWithDiagnostics(options);
-  return result.extensions;
-}
-
-async function discoverGovernanceExtensionsWithDiagnostics(
-  options: DiscoverGovernanceExtensionsOptions = {}
-): Promise<{
-  extensions: DiscoveredGovernanceExtension[];
-  diagnostics: GovernanceExtensionDiagnostic[];
-}> {
-  const moduleLoader = options.moduleLoader ?? defaultGovernanceModuleLoader;
-  const extensions: DiscoveredGovernanceExtension[] = [];
-  const diagnostics: GovernanceExtensionDiagnostic[] = [];
-  const loadRequests = [...(options.loadRequests ?? [])];
-
-  if (loadRequests.some((request) => request.source === 'legacy')) {
-    diagnostics.push({
-      code: 'governance.extension.legacy_probing_used',
-      severity: 'warning',
-      message:
-        'Legacy governance extension probing from nx.json.plugins is deprecated. Register governance extensions explicitly under nx.json.governance.extensions instead.',
-      legacy: true,
-    });
-  }
-
-  for (const loadRequest of loadRequests) {
-    try {
-      const loadedModule = await moduleLoader(loadRequest.moduleSpecifier);
-      const governanceExtension =
-        readGovernanceExtensionDefinition(loadedModule);
-
-      extensions.push({
-        sourceSpecifier: loadRequest.sourceSpecifier,
-        moduleSpecifier: loadRequest.moduleSpecifier,
-        source: loadRequest.source,
-        optional: loadRequest.optional,
-        definition: governanceExtension,
-      });
-      diagnostics.push({
-        code: 'governance.extension.loaded',
-        severity: 'notice',
-        message: `Loaded governance extension from "${loadRequest.moduleSpecifier}".`,
-        packageName: loadRequest.sourceSpecifier,
-        moduleSpecifier: loadRequest.moduleSpecifier,
-        extensionId: governanceExtension.id,
-        legacy: loadRequest.source === 'legacy',
-      });
-    } catch (error) {
-      if (
-        shouldSkipMissingGovernanceExtension(
-          error,
-          loadRequest.moduleSpecifier,
-          loadRequest
-        )
-      ) {
-        diagnostics.push(
-          createMissingExtensionDiagnostic(
-            loadRequest,
-            loadRequest.source === 'legacy'
-              ? 'governance.extension.legacy_entrypoint_missing'
-              : 'governance.extension.skipped_optional_missing'
-          )
-        );
-        continue;
-      }
-
-      if (
-        loadRequest.source === 'explicit' &&
-        isMissingDirectModuleLookup(error, loadRequest.moduleSpecifier)
-      ) {
-        diagnostics.push(
-          createMissingExtensionDiagnostic(
-            loadRequest,
-            'governance.extension.missing_required'
-          )
-        );
-        throw new GovernanceExtensionRegistrationError(
-          toErrorMessage(error),
-          diagnostics
-        );
-      }
-
-      if (isInvalidGovernanceExtensionDefinitionError(error)) {
-        diagnostics.push({
-          code: 'governance.extension.invalid_definition',
-          severity: 'error',
-          message: toErrorMessage(error),
-          packageName: loadRequest.sourceSpecifier,
-          moduleSpecifier: loadRequest.moduleSpecifier,
-          legacy: loadRequest.source === 'legacy',
-        });
-        throw new GovernanceExtensionRegistrationError(
-          toErrorMessage(error),
-          diagnostics
-        );
-      }
-
-      throw error;
-    }
-  }
-
-  return {
-    extensions,
-    diagnostics,
-  };
-}
-
-export async function registerGovernanceExtensions(
-  context: GovernanceExtensionHostContext,
-  options: RegisterGovernanceExtensionsOptions = {}
-): Promise<GovernanceExtensionRegistry> {
-  const result = await registerGovernanceExtensionsWithDiagnostics(
-    context,
-    options
-  );
-  return result.registry;
-}
-
-export async function registerGovernanceExtensionsWithDiagnostics(
-  context: GovernanceExtensionHostContext,
-  options: RegisterGovernanceExtensionsOptions = {}
-): Promise<GovernanceExtensionRegistrationResult> {
-  const registry: GovernanceExtensionRegistry = {
-    metricProviders: [],
-    signalProviders: [],
-    rulePacks: [],
-    enrichers: [],
-  };
-  const discoveryResult = await discoverGovernanceExtensionsWithDiagnostics(
-    options
-  );
-  const diagnostics = [...discoveryResult.diagnostics];
-  const discoveredExtensions = discoveryResult.extensions;
-  const seenExtensionIds = new Map<string, string>();
-
-  for (const extension of discoveredExtensions) {
-    try {
-      validateGovernanceExtensionId(extension);
-    } catch (error) {
-      diagnostics.push({
-        code: 'governance.extension.invalid_definition',
-        severity: 'error',
-        message: toErrorMessage(error),
-        packageName: extension.sourceSpecifier,
-        moduleSpecifier: extension.moduleSpecifier,
-        legacy: extension.source === 'legacy',
-      });
-      throw new GovernanceExtensionRegistrationError(
-        toErrorMessage(error),
-        diagnostics
-      );
-    }
-
-    const previousModule = seenExtensionIds.get(extension.definition.id);
-    if (previousModule) {
-      const message = `Duplicate governance extension id "${extension.definition.id}" was found in "${previousModule}" and "${extension.moduleSpecifier}".`;
-      diagnostics.push({
-        code: 'governance.extension.duplicate_id',
-        severity: 'error',
-        message,
-        packageName: extension.sourceSpecifier,
-        moduleSpecifier: extension.moduleSpecifier,
-        extensionId: extension.definition.id,
-        legacy: extension.source === 'legacy',
-      });
-      throw new GovernanceExtensionRegistrationError(message, diagnostics);
-    }
-
-    seenExtensionIds.set(extension.definition.id, extension.moduleSpecifier);
-
-    try {
-      const host = new GovernanceExtensionHost(
-        context,
-        extension.definition.id
-      );
-      await extension.definition.register(host);
-      mergeRegistry(registry, host.toRegistry());
-    } catch (error) {
-      const message = `Governance extension "${
-        extension.definition.id
-      }" from "${
-        extension.moduleSpecifier
-      }" failed during registration: ${toErrorMessage(error)}`;
-      diagnostics.push({
-        code: 'governance.extension.registration_failed',
-        severity: 'error',
-        message,
-        packageName: extension.sourceSpecifier,
-        moduleSpecifier: extension.moduleSpecifier,
-        extensionId: extension.definition.id,
-        legacy: extension.source === 'legacy',
-      });
-      throw new GovernanceExtensionRegistrationError(message, diagnostics);
-    }
-  }
-
-  return {
-    registry,
-    diagnostics,
-  };
-}
-
-export async function applyGovernanceEnrichers(
-  registry: GovernanceExtensionRegistry,
-  input: GovernanceWorkspaceEnricherInput
-): Promise<GovernanceWorkspace> {
-  let workspace = input.workspace;
-
-  for (const enricher of registry.enrichers) {
-    const enrichedWorkspace = await enricher.contribution.enrichWorkspace({
-      ...input,
-      workspace,
-    });
-    workspace = enrichedWorkspace;
-  }
-
-  return workspace;
-}
-
-export async function evaluateGovernanceRulePacks(
-  registry: GovernanceExtensionRegistry,
-  input: GovernanceRulePackInput
-): Promise<Violation[]> {
-  const violations = await Promise.all(
-    registry.rulePacks.map(async (rulePack) => {
-      const providedViolations = await rulePack.contribution.evaluate(input);
-      return providedViolations.map((violation) => ({
-        ...violation,
-        sourcePluginId: violation.sourcePluginId ?? rulePack.pluginId,
-      }));
-    })
-  );
-
-  return violations.flat();
-}
-
-export async function collectGovernanceSignals(
-  registry: GovernanceExtensionRegistry,
-  input: GovernanceSignalProviderInput
-): Promise<GovernanceSignal[]> {
-  const signals = await Promise.all(
-    registry.signalProviders.map(async (signalProvider) => {
-      const providedSignals = await signalProvider.contribution.provideSignals(
-        input
-      );
-      return providedSignals.map((signal) => ({
-        ...signal,
-        source: 'extension' as const,
-        sourcePluginId: signal.sourcePluginId ?? signalProvider.pluginId,
-      }));
-    })
-  );
-
-  return signals.flat();
-}
-
-export async function collectGovernanceMeasurements(
-  registry: GovernanceExtensionRegistry,
-  input: GovernanceMetricProviderInput
-): Promise<Measurement[]> {
-  const measurements = await Promise.all(
-    registry.metricProviders.map(async (metricProvider) => {
-      const providedMeasurements =
-        await metricProvider.contribution.provideMetrics(input);
-      return providedMeasurements.map((measurement) => ({
-        ...measurement,
-        sourcePluginId: measurement.sourcePluginId ?? metricProvider.pluginId,
-      }));
-    })
-  );
-
-  return measurements.flat();
-}
-
-function mergeRegistry(
-  target: GovernanceExtensionRegistry,
-  source: GovernanceExtensionRegistry
-): void {
-  target.metricProviders.push(...source.metricProviders);
-  target.signalProviders.push(...source.signalProviders);
-  target.rulePacks.push(...source.rulePacks);
-  target.enrichers.push(...source.enrichers);
-}
-
-function createMissingExtensionDiagnostic(
-  loadRequest: GovernanceExtensionLoadRequest,
-  code:
-    | 'governance.extension.skipped_optional_missing'
-    | 'governance.extension.missing_required'
-    | 'governance.extension.legacy_entrypoint_missing'
-): GovernanceExtensionDiagnostic {
-  const message =
-    code === 'governance.extension.skipped_optional_missing'
-      ? `Skipped optional governance extension "${loadRequest.moduleSpecifier}" because the package could not be resolved.`
-      : code === 'governance.extension.missing_required'
-      ? `Required governance extension "${loadRequest.moduleSpecifier}" could not be resolved.`
-      : `Skipped legacy governance extension probe for "${loadRequest.moduleSpecifier}" because the governance entrypoint is missing.`;
-
-  return {
-    code,
-    severity:
-      code === 'governance.extension.missing_required' ? 'error' : 'notice',
-    message,
-    packageName: loadRequest.sourceSpecifier,
-    moduleSpecifier: loadRequest.moduleSpecifier,
-    legacy: loadRequest.source === 'legacy',
-  };
-}
-
-function readGovernanceExtensionDefinition(
-  loadedModule: unknown
-): GovernanceExtensionDefinition {
-  const governanceExtension = (
-    loadedModule as { governanceExtension?: unknown }
-  )?.governanceExtension;
-
-  if (
-    !governanceExtension ||
-    typeof governanceExtension !== 'object' ||
-    typeof (governanceExtension as GovernanceExtensionDefinition).register !==
-      'function'
-  ) {
-    throw new Error(
-      'Governance extension module must export a named "governanceExtension" definition.'
-    );
-  }
-
-  return governanceExtension as GovernanceExtensionDefinition;
-}
-
-function validateGovernanceExtensionId(
-  extension: DiscoveredGovernanceExtension
-): void {
-  if (
-    typeof extension.definition.id !== 'string' ||
-    extension.definition.id.trim().length === 0
-  ) {
-    throw new Error(
-      `Governance extension module "${extension.moduleSpecifier}" must declare a non-empty "id".`
-    );
-  }
-}
-
-async function defaultGovernanceModuleLoader(
-  specifier: string
-): Promise<unknown> {
-  return import(specifier);
-}
-
-function isMissingGovernanceEntrypoint(
-  error: unknown,
-  moduleSpecifier: string
-): boolean {
-  const errorRecord =
-    error && typeof error === 'object'
-      ? (error as { code?: string; message?: string })
-      : undefined;
-
-  const message = errorRecord?.message;
-  if (typeof message !== 'string') {
-    return false;
-  }
-
-  const errorCode = errorRecord?.code;
-  if (errorCode === 'ERR_PACKAGE_PATH_NOT_EXPORTED') {
-    return matchesGovernanceEntrypointSubpath(message, moduleSpecifier);
-  }
-
-  if (
-    errorCode === 'ERR_MODULE_NOT_FOUND' ||
-    errorCode === 'MODULE_NOT_FOUND'
-  ) {
-    return matchesGovernanceEntrypointLookup(message, moduleSpecifier);
-  }
-
-  return (
-    message.startsWith('Cannot find module') &&
-    matchesGovernanceEntrypointLookup(message, moduleSpecifier)
-  );
-}
-
-function shouldSkipMissingGovernanceExtension(
-  error: unknown,
-  moduleSpecifier: string,
-  loadRequest: GovernanceExtensionLoadRequest
-): boolean {
-  if (loadRequest.source === 'legacy') {
-    return isMissingGovernanceEntrypoint(error, moduleSpecifier);
-  }
-
-  return (
-    loadRequest.optional === true &&
-    isMissingDirectModuleLookup(error, moduleSpecifier)
-  );
-}
-
-function isMissingDirectModuleLookup(
-  error: unknown,
-  moduleSpecifier: string
-): boolean {
-  const errorRecord =
-    error && typeof error === 'object'
-      ? (error as { code?: string; message?: string })
-      : undefined;
-
-  const message = errorRecord?.message;
-  if (typeof message !== 'string') {
-    return false;
-  }
-
-  const errorCode = errorRecord?.code;
-  if (
-    errorCode === 'ERR_MODULE_NOT_FOUND' ||
-    errorCode === 'MODULE_NOT_FOUND'
-  ) {
-    return matchesGovernanceEntrypointLookup(message, moduleSpecifier);
-  }
-
-  return (
-    message.startsWith('Cannot find module') &&
-    matchesGovernanceEntrypointLookup(message, moduleSpecifier)
-  );
-}
-
-function isInvalidGovernanceExtensionDefinitionError(error: unknown): boolean {
-  return (
-    error instanceof Error &&
-    error.message ===
-      'Governance extension module must export a named "governanceExtension" definition.'
-  );
-}
-
-function matchesGovernanceEntrypointSubpath(
-  message: string,
-  moduleSpecifier: string
-): boolean {
-  const { packageName, subpath } = splitPackageSubpath(moduleSpecifier);
-
-  return (
-    message.includes('Package subpath') &&
-    message.includes(packageName) &&
-    message.includes(subpath)
-  );
-}
-
-function matchesGovernanceEntrypointLookup(
-  message: string,
-  moduleSpecifier: string
-): boolean {
-  const normalizedSpecifier = normalizeLookupTarget(moduleSpecifier);
-  const quotedTargets = extractQuotedValues(message);
-
-  return quotedTargets.some((target) =>
-    isMatchingGovernanceLookupTarget(target, normalizedSpecifier)
-  );
-}
-
-function isMatchingGovernanceLookupTarget(
-  target: string,
-  normalizedSpecifier: string
-): boolean {
-  const normalizedTarget = normalizeLookupTarget(target);
-
-  return (
-    normalizedTarget === normalizedSpecifier ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}.js`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}.mjs`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}.cjs`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}/index.js`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}/index.mjs`) ||
-    normalizedTarget.endsWith(`/${normalizedSpecifier}/index.cjs`)
-  );
-}
-
-function extractQuotedValues(message: string): string[] {
-  return [...message.matchAll(/['"]([^'"]+)['"]/g)].map(
-    (match) => match[1] ?? ''
-  );
-}
-
-function normalizeLookupTarget(target: string): string {
-  return target
-    .replace(/^file:\/\/\/?/, '/')
-    .replaceAll('\\', '/')
-    .replace(/\/+/g, '/')
-    .replace(/\/$/, '');
-}
-
-function splitPackageSubpath(moduleSpecifier: string): {
-  packageName: string;
-  subpath: string;
-} {
-  const segments = moduleSpecifier.split('/');
-  const packageSegments = moduleSpecifier.startsWith('@')
-    ? segments.slice(0, 2)
-    : segments.slice(0, 1);
-  const packageName = packageSegments.join('/');
-  const subpathSegments = segments.slice(packageSegments.length);
-
-  return {
-    packageName,
-    subpath: `./${subpathSegments.join('/')}`,
-  };
-}
-
-function toErrorMessage(error: unknown): string {
-  return error instanceof Error ? error.message : 'Unknown error.';
-}
+export * from './runtime.js';
